requirements.txt
flask==2.0.1
flask-cors==3.0.10
google-cloud-aiplatform>=1.36.0
gunicorn==20.1.0